Fixed: error - the CustomerService was inside another CustomerSignupService.java class instead of the service folder.
