--- src/main/java/com/chameleon/estaciona_uai/dto/CustomerSignupDto.java
@@ -1,14 +1,0 @@
-package com.chameleon.estaciona_uai.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CustomerSignupDto {
-    private String name;
-    private String email;
-    private String password;
-}
